Правки по Отрисуй меня полностью (часть 2)
