+++ js/main.js
@@ -4,7 +4,6 @@ import './form.js';
 import './filter.js';
 import './message.js';
 import './api.js';
-import './input-form.js';
